pyproject.toml
[project]
name = "dioxide-python-sdk"
version = "0.6.0"
description = "dioxide python SDK"
authors = [
    {name = "phenix3443",email = "phenix3443@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crc32c==2.7.post1",
    "krock32==0.1.1",
    "python-box==7.2.0",
    "requests==2.32.3",
    "setuptools==75.6.0",
    "websockets==11.0.3",
    "pynacl (>=1.6.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "<8.0.0"

